requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
line-bot-sdk==3.12.0
langchain==0.2.6
langchain-community==0.2.6
chromadb==0.5.5
pypdf==4.3.1
ollama==0.2.1
